sync fork
the fork `planemo-autoupdate/tools-iuc` is currently
not synced.

so the for loop will include deprecated tools (and exclude
new tools). For these the `git checkout -b $REPO upstream/master`
will use the newest upstream branch where the tool dir is absent
leading to an error.

in the future we might try to deleted branches and PR
if tools are deprecated